// hold/EdFi.Buzz.UI/src/components/Survey/types/SurveyQuestionType.ts
// SPDX-License-Identifier: Apache-2.0
// Licensed to the Ed-Fi Alliance under one or more agreements.
// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
// See the LICENSE and NOTICES files in the project root for more information.

export type SurveyQuestionType = {
  id: string;
  question: string;
  answer: string;
  comments?: string;
  disabled?: boolean;
};